Добавлена дополнительная безопасность в сервисы

+++ src/main/java/ru/skypro/homework/service/impl/CommentServiceImpl.java
@@ -2,13 +2,16 @@ package ru.skypro.homework.service.impl;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Component;
-import org.springframework.stereotype.Service;
 import ru.skypro.homework.dto.Comment;
 import ru.skypro.homework.dto.Comments;
 import ru.skypro.homework.dto.CreateOrUpdateComment;
+import ru.skypro.homework.dto.Role;
+import ru.skypro.homework.exception.AdNotFoundException;
+import ru.skypro.homework.exception.CommentNotFoundException;
+import ru.skypro.homework.exception.ForbiddenActionException;
+import ru.skypro.homework.mappers.AdMapper;
 import ru.skypro.homework.mappers.CommentMapper;
 import ru.skypro.homework.model.AdEntity;
 import ru.skypro.homework.model.CommentEntity;
@@ -18,6 +21,9 @@ import ru.skypro.homework.repository.CommentRepository;
 import ru.skypro.homework.repository.UserRepository;
 import ru.skypro.homework.service.CommentService;
 
+import java.util.List;
+import java.util.stream.Collectors;
+
 @Component
 @RequiredArgsConstructor
 @Slf4j
@@ -26,12 +32,18 @@ public class CommentServiceImpl implements CommentService {
     private final AdRepository adRepository;
     private final CommentRepository commentRepository;
     private final UserRepository userRepository;
-    private CommentMapper mapper;
+    private CommentMapper commentMapper;
+    private AdMapper adMapper;
 
     @Override
     public Comments getCommentsAdv(Integer id) {
-        AdEntity adEntity = adRepository.findAdEntityByPk(id);
-        return new Comments(adEntity.getComments().size(), (Comment[]) adEntity.getComments().toArray());
+        AdEntity adEntity = adRepository.findById(id).orElseThrow(() -> new AdNotFoundException("Ошибка! Реклама с данным id не найдена"));
+        List<Comment> comments = adEntity.getComments().stream().map(this::mapCommentEntityToComment).collect(Collectors.toList());
+        return new Comments(comments.size(), comments.toArray(new Comment[0]));
+    }
+
+    private Comment mapCommentEntityToComment(CommentEntity commentEntity) {
+        return commentMapper.mapCommentEntityToComment(commentEntity);
     }
 
     @Override
@@ -39,20 +51,30 @@ public class CommentServiceImpl implements CommentService {
         UserEntity user = userRepository.findByEmail(username).orElseThrow(() -> new UsernameNotFoundException("Ошибка! Пользователь не найден!"));
         AdEntity ad = adRepository.findAdEntityByPk(id);
         CommentEntity comment = new CommentEntity(user.getImage(), user.getFirstName(), id, createOrUpdateComment.getText(), ad, user);
-        return mapper.mapCommentEntityToComment(commentRepository.save(comment));
+        return commentMapper.mapCommentEntityToComment(commentRepository.save(comment));
     }
 
     @Override
     public void deleteCommentAdv(Integer adId, Integer commentId, String username) {
+        CommentEntity comment = commentRepository.findById(adId).orElseThrow(() -> new CommentNotFoundException("Ошибка! Комментарий с указанным ID не найден"));
+        boolean isAuthor = username.equals(comment.getAuthor().getEmail());
+        boolean isAdmin = comment.getAuthor().getRole().equals(Role.ADMIN);
+        if (!isAuthor && !isAdmin) {
+            throw new ForbiddenActionException("Пользователь " + username + " не имеет прав удалять чужие комментарии.");
+        }
         commentRepository.deleteById(commentId);
     }
 
     @Override
     public Comment updateCommentAdv(Integer adId, Integer commentId, CreateOrUpdateComment createOrUpdateComment, String username) {
-        UserEntity user = userRepository.findByEmail(username).orElseThrow(() -> new UsernameNotFoundException("Ошибка! Пользователь не найден!"));
-        AdEntity ad = adRepository.findAdEntityByPk(adId);
-        CommentEntity comment = new CommentEntity(user.getImage(), user.getFirstName(), commentId, createOrUpdateComment.getText(), ad, user);
+        CommentEntity comment = commentRepository.findById(commentId).orElseThrow(() -> new CommentNotFoundException("Ошибка! Комментарий с указанным ID не найден"));
+        boolean isAuthor = username.equals(comment.getAuthor().getEmail());
+        boolean isAdmin = comment.getAuthor().getRole().equals(Role.ADMIN);
+        if (!isAuthor && !isAdmin) {
+            throw new ForbiddenActionException("Пользователь " + username + " не имеет прав удалять чужие комментарии.");
+        }
+        comment.setText(createOrUpdateComment.getText());
         commentRepository.save(comment);
-        return mapper.mapCommentEntityToComment(comment);
+        return commentMapper.mapCommentEntityToComment(comment);
     }
 }
